Update README.md
scroll down to see the images instead of them being stuck at the top. Also, you want the second page to be on the same page but further down, not a separate file.

Here’s how we can fix it:

Make the images appear lower on the page instead of being at the top.

Create a button that smoothly scrolls down to the second section.

Style it so it looks clean and cute!